feat(#636): fix jtcop suggestions
